+++ lib/snorby/jobs/mass_classification.rb
@@ -18,11 +18,10 @@
 
 module Snorby
   module Jobs
-    class MassClassification < Struct.new(:classification_id, 
-                                          :options, :user_id, :reclassify)
+    class MassClassification < Struct.new(:ids, :classification_id, :user_id)
 
       def perform
-        Event.update_classification(ids, params[:classification_id], User.current_user.id)
+        Event.update_classification(ids, classification_id.to_i, user_id.to_i)
       end
 
     end